// packages/app/src/types/redux.ts
export type Message = {
    _id: string;
    type: string;
    content: string;
    createTime: number;
    percent?: number;
    loading?: boolean;
    from: {
        _id: string;
        username: string;
        avatar: string;
        tag: string;
        originUsername?: string;
    };
    to: string;
    deleted?: boolean;
};

export type Group = {
    _id: string;
    type: 'group';
    name: string;
    avatar: string;
    messages: Message[];
    unread: number;
    members: {
        _id: string;
        user: {
            _id: string;
            username: string;
            avatar: string;
        };
        os: string;
        browser: string;
        environment: string;
    }[];
    creator: string;
    createTime: number;
};

export type Friend = {
    _id: string;
    type: 'friend';
    name: string;
    avatar: string;
    from: string;
    to: {
        _id: string;
        avatar: string;
        username: string;
    };
    messages: Message[];
    unread: number;
    createTime: number;
    isOnline?: boolean;
};

export type Temporary = {
    _id: string;
    type: 'temporary';
    name: string;
    avatar: string;
    messages: Message[];
    unread: number;
    createTime: number;
    isOnline?: boolean;
};

export type Linkman = Group | Friend | Temporary;